fix: read errors as an array in range spinbox
