+++ src/components/Form/RangeSpinBox.tsx
@@ -5,6 +5,8 @@ import _ from 'lodash'
 import { Field, FormikErrors, FormikTouched } from 'formik'
 import { InputGroup } from 'reactstrap'
 
+import { getFormikError } from '../../helpers/getFormikError'
+
 export interface FormSpinBoxProps<T> {
   identifier: string
   step?: number | string
@@ -16,15 +18,15 @@ export interface FormSpinBoxProps<T> {
 }
 
 export function RangeSpinBox<T>({ identifier, step, min, max, pattern, errors, touched }: FormSpinBoxProps<T>) {
-  const isTouched = _.get(touched, identifier)
-  const errorMessage = _.get(errors, identifier)
-  const showError = errorMessage && isTouched
-  const borderDanger = showError ? 'border-danger' : ''
+  const errorFirst = getFormikError({ errors, touched, identifier: `${identifier}[0]` })
+  const errorSecond = getFormikError({ errors, touched, identifier: `${identifier}[1]` })
+  const borderDangerFirst = errorFirst ? 'border-danger' : ''
+  const borderDangerSecond = errorSecond ? 'border-danger' : ''
 
   return (
     <InputGroup>
       <Field
-        className={`form-control d-inline ${borderDanger}`}
+        className={`form-control d-inline ${borderDangerFirst}`}
         id={identifier}
         name={`${identifier}[0]`}
         type="number"
@@ -35,7 +37,7 @@ export function RangeSpinBox<T>({ identifier, step, min, max, pattern, errors, t
       />
       <span className="h-100 pt-2 px-1 text-bold">{'-'}</span>
       <Field
-        className={`form-control d-inline  ${borderDanger}`}
+        className={`form-control d-inline  ${borderDangerSecond}`}
         id={identifier}
         name={`${identifier}[1]`}
         type="number"
